% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_signal_plots.R
\name{assemble_heatmap_cluster_bars}
\alias{assemble_heatmap_cluster_bars}
\title{assemble_heatmap_cluster_bars}
\usage{
assemble_heatmap_cluster_bars(plots, ...)
}
\arguments{
\item{plots}{list of plots as returned from ssvSignalHeatmap.ClusterBars when return_unassembled_plots = TRUE}

\item{...}{arguments passed to cowplot::plot_grid}
}
\value{
A grob produced by cowplot::plot_grid
}
\description{
assemble_heatmap_cluster_bars
}
\examples{
data(CTCF_in_10a_profiles_gr)
plots = ssvSignalHeatmap.ClusterBars(CTCF_in_10a_profiles_gr, return_unassembled_plots = TRUE)
assemble_heatmap_cluster_bars(plots)
}
