% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_CTCF.R
\docType{data}
\name{Bcell_peaks}
\alias{Bcell_peaks}
\title{4 random peaks for paired-end data}
\format{
GRanges length 4
}
\description{
matches \code{system.file("extdata/Bcell_PE.mm10.bam", package = "seqsetvis")}
}
\details{
this is included only for testing ssvFetchBamPE functions.
}
\keyword{datasets}
