% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_chromHMM.R
\docType{data}
\name{chromHMM_demo_bw_states_gr}
\alias{chromHMM_demo_bw_states_gr}
\title{MCF10A CTCF profiles at 20 windows per chromHMM state, hg38.}
\format{
a GRanges object of length 4000
with 5 metadata columns sufficient for use with ggplot2
}
\value{
a GRanges object of length 4000
with 5 metadata columns sufficient for use with ggplot2
}
\description{
MCF10A CTCF profiles at 20 windows per chromHMM state, hg38.
}
\details{
part of \code{\link{chromHMM_demo_data}}

the result of ssvFetchBigwig() on the MCF10A_CTCF_FE.bw
near 20 randomly selected windows per chromHMM state.
}
\keyword{datasets}
