% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_chromHMM.R
\docType{data}
\name{chromHMM_demo_chain_url}
\alias{chromHMM_demo_chain_url}
\title{URL to download hg19ToHg38 liftover chain from UCSC}
\format{a character containing a URL}
\description{
URL to download hg19ToHg38 liftover chain from UCSC
}
\details{
file is gzipped .txt

part of \code{\link{chromHMM_demo_data}}
}
\keyword{datasets}
