% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_chromHMM.R
\docType{data}
\name{chromHMM_demo_data}
\alias{chromHMM_demo_data}
\title{chromHMM state segmentation in the MCF7 cell line}
\description{
Vignette data for seqsetvis was downloaded directly from GEO series \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE57498}{GSE57498}.
This data is the state segmentation by chromHMM in the MCF7 cell line.
chromHMM creates a hidden markov model by integrating several ChIP-seq
samples, in this case: \itemize{
\item MCF7_H3K27ac_ChIP-Seq
\item MCF7_H3K27me3_ChIP-Seq
\item MCF7_H3K4me1_ChIP-Seq
\item MCF7_H3K4me3_ChIP-Seq
\item MCF7_RNApolIIp_ChIP-Seq
}

Data from GEO series \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE57498}{GSE57498}
is from the publication \href{https://www.ncbi.nlm.nih.gov/pubmed/24916973}{Taberlay PC et al. 2014}
}
\details{
Contains: \itemize{
\item \code{\link{chromHMM_demo_overlaps_gr}}
\item \code{\link{chromHMM_demo_bw_states_gr}}
\item \code{\link{chromHMM_demo_state_colors}}
\item \code{\link{chromHMM_demo_segmentation_url}}
\item \code{\link{chromHMM_demo_chain_url}}
}
}
\keyword{datasets}
