% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_chromHMM.R
\docType{data}
\name{chromHMM_demo_overlaps_gr}
\alias{chromHMM_demo_overlaps_gr}
\title{overlap of MCF10A CTCF with MCF7 chromHMM states, hg38.}
\format{a GRanges object of length 98
with 10 logical metadata columns, 1 per state.}
\description{
overlap of MCF10A CTCF with MCF7 chromHMM states, hg38.
}
\details{
part of \code{\link{chromHMM_demo_data}}

the result of ssvOverlapIntervalSets() on MCF10A CTCF peaks and
         MCF7 chromHMM states with use_first = TRUE

first (the MCF10A peaks) and no_hit columns have been removed
each remaining column represents MCF10A peaks overlapping with a state.
}
\keyword{datasets}
