% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_chromHMM.R
\docType{data}
\name{chromHMM_demo_segmentation_url}
\alias{chromHMM_demo_segmentation_url}
\title{URL to download hg19 MCF7 chromHMM segmentation}
\format{
a character containing a URL
}
\description{
URL to download hg19 MCF7 chromHMM segmentation
}
\details{
file is gzipped bed with name, score, itemRgb and thick meta columns

part of \code{\link{chromHMM_demo_data}}
}
\keyword{datasets}
