% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_chromHMM.R
\docType{data}
\name{chromHMM_demo_state_total_widths}
\alias{chromHMM_demo_state_total_widths}
\title{state name to total width mappings, hg38}
\format{
named numeric of total widths per state
}
\value{
named numeric of total widths per state
}
\description{
state name to total width mappings, hg38
}
\details{
part of \code{\link{chromHMM_demo_data}}
}
\keyword{datasets}
