% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_signal_operations.R
\name{clusteringKmeansNestedHclust}
\alias{clusteringKmeansNestedHclust}
\title{perform kmeans clustering on matrix rows and return reordered matrix along with order matched cluster assignments
clusters are sorted using hclust on centers
the contents of each cluster are sorted using hclust}
\usage{
clusteringKmeansNestedHclust(mat, nclust, seed = 0)
}
\arguments{
\item{mat}{A wide format matrix}

\item{nclust}{the number of clusters}

\item{seed}{passed to set.seed() to allow reproducibility}
}
\description{
perform kmeans clustering on matrix rows and return reordered matrix along with order matched cluster assignments
clusters are sorted using hclust on centers
the contents of each cluster are sorted using hclust
}
\examples{
dt = data.table::copy(CTCF_in_10a_profiles_dt)
mat = data.table::dcast(dt, id ~ sample + x, value.var = "FE" )
rn = mat$id
mat = as.matrix(mat[,-1])
rownames(mat) = rn
clust_dt = clusteringKmeansNestedHclust(mat, nclust = 3)
dt = merge(dt, clust_dt)
dt$id = factor(dt$id, levels = clust_dt$id)
dt[order(id)]
}
