% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{col2hex}
\alias{col2hex}
\title{converts a valid r color name ("black", "red", "white", etc.) to a hex value}
\usage{
col2hex(color_name)
}
\arguments{
\item{color_name}{character. one or more r color names.}
}
\value{
hex value of colors coded by colors()
}
\description{
converts a valid r color name ("black", "red", "white", etc.) to a hex value
}
\examples{
col2hex(c("red", "green", "blue"))
col2hex(c("lightgray", "gray", "darkgray"))
}
