% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_CTCF.R
\docType{data}
\name{CTCF_in_10a_bigWig_urls}
\alias{CTCF_in_10a_bigWig_urls}
\title{FTP URL path for vignette data.}
\format{
named character vector of length 3
}
\description{
FE bigWig tracks for CTCF ChIP-seq in a MCF10A progression model.
See GEO series GSE98551 for details.
}
\details{
part of \code{\link{CTCF_in_10a_data}}
}
\keyword{datasets}
