% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{CTCF_in_10a_profiles_dt}
\alias{CTCF_in_10a_profiles_dt}
\alias{CTCF_in_10a_overlaps_gr}
\alias{CTCF_in_10a_bigWig_urls}
\alias{CTCF_in_10a_narrowPeak_urls}
\title{Profiles for 100 randomly selected regions from overlapping CTCF peaks in 10a cell ChIP-seq
Results from CTCF_in_10a_overlaps_gr}
\format{A data frame with 53940 rows and 10 variables}
\description{
A tidy data.table at window size 100 bp within 300 bp of peak center
 The variables are as follows:

macs2 peak calls for CTCF ChIP-seq in a MCF10A progression model.
See  GEO series GSE98551 for details.

FE bigWig tracks for CTCF ChIP-seq in a MCF10A progression model.
See GEO series GSE98551 for details.

macs2 peak calls for CTCF ChIP-seq in a MCF10A progression model.
See  GEO series GSE98551 for details.
}
\details{
\itemize{
  \item seqnames. chromosome for GRanges compatibility
  \item start. start of interval
  \item end. end of interval
  \item id. unique identifier
  \item y. fold-enrichment over input.
  \item x. bp relative to center
  \item sample. name of originating sample
}
}
\keyword{datasets}
