% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_CTCF.R
\docType{data}
\name{CTCF_in_10a_narrowPeak_grs}
\alias{CTCF_in_10a_narrowPeak_grs}
\title{list of GRanges that results in 100 random subset when overlapped}
\format{
named character vector of length 3
}
\description{
list of GRanges that results in 100 random subset when overlapped
}
\details{
part of \code{\link{CTCF_in_10a_data}}
}
\keyword{datasets}
