% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{CTCF_in_10a_overlaps_gr}
\alias{CTCF_in_10a_overlaps_gr}
\title{100 randomly selected regions from overlapping CTCF peaks in 10a cell ChIP-seq}
\format{GenomicRanges with 3 metadata columns of membership table}
\description{
macs2 peak calls for CTCF ChIP-seq in a MCF10A progression model.
See  GEO series GSE98551 for details.
}
\keyword{datasets}
