% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{CTCF_in_10a_profiles_dt}
\alias{CTCF_in_10a_profiles_dt}
\title{100 randomly selected regions from overlapping CTCF peaks in 10a cell ChIP-seq}
\format{A data frame with 53940 rows and 10 variables}
\usage{
data(CTCF_in_10a_profiles_dt)
}
\description{
A tidy data.table at window size 100 bp within 300 bp of peak center
 The variables are as follows:
}
\details{
\itemize{
  \item seqnames. chromosome for GRanges compatibility
  \item start. start of interval
  \item end. end of interval
  \item id. unique identifier
  \item FE. fold-enrichment over input.
  \item x. bp relative to center
  \item sample. name of originating sample
}
}
\keyword{datasets}
