% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fetch_bam.R
\name{.expand_cigar_dt}
\alias{.expand_cigar_dt}
\title{Expand intermediate bam fetch by cigar codes}
\usage{
.expand_cigar_dt(cigar_dt, op_2count = c("M", "D", "=", "X"))
}
\arguments{
\item{cigar_dt}{data.table with 5 required named columns in any order.
c("which_label", "seqnames", "strand", "start", "cigar")}

\item{op_2count}{Cigar codes to count. Default is alignment (M), deletion
(D), match (=), and mismatch (X).  Other useful codes may be skipped
regions for RNA splicing (N).  The locations of any insterions (I) or
clipping/padding (S, H, or P) will be a single bp immediately before the
interval.}
}
\value{
data.table with cigar entries expanded
}
\description{
see \href{https://samtools.github.io/hts-specs/SAMv1.pdf}{sam specs} for
cigar details
}
