% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fetch_bam.R
\name{.expand_cigar_dt_recursive}
\alias{.expand_cigar_dt_recursive}
\title{Expand intermediate bam fetch by cigar codes}
\usage{
.expand_cigar_dt_recursive(cigar_dt)
}
\arguments{
\item{cigar_dt}{data.table with 5 required named columns in any order.
c("which_label", "seqnames", "strand", "start", "cigar")}
}
\value{
data.table with cigar entries expanded
}
\description{
see \href{https://samtools.github.io/hts-specs/SAMv1.pdf}{sam specs}
for cigar details
}
