% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fetch_bam.R
\name{.rm_dupes}
\alias{.rm_dupes}
\title{Remove duplicate reads based on stranded start position.  This is an
over-simplification.  For better duplicate handling, duplicates must be
marked in bam and flag passed to fetchBam() ... for ScanBamParam}
\usage{
.rm_dupes(reads_dt, max_dupes)
}
\arguments{
\item{reads_dt}{data.table of reads as loaded by fetchBam}

\item{max_dupes}{maximum allowed positional duplicates}
}
\value{
reads_dt with duplicated reads over max_dupes removed
}
\description{
flag = scanBamFlag(isDuplicate = FALSE)
}
