% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fetch_bamPE.R
\name{.rm_dupesPE}
\alias{.rm_dupesPE}
\title{Remove duplicate paired-end reads based on start and end position.  This is
an over-simplification.  For better duplicate handling, duplicates must be
marked in bam and flag passed to fetchBamPE() ... for ScanBamParam}
\usage{
.rm_dupesPE(reads_dt, max_dupes)
}
\arguments{
\item{reads_dt}{data.table of reads as loaded by fetchBamPE}

\item{max_dupes}{maximum allowed positional duplicates}
}
\value{
reads_dt with duplicated reads over max_dupes removed
}
\description{
flag = scanBamFlag(isDuplicate = FALSE)
}
