% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_easyLoad.R
\name{easyLoad_narrowPeak}
\alias{easyLoad_narrowPeak}
\title{easyLoad_narrowPeak takes a character vector of file paths to narrowPeak
files from MACS2 and returns a named list of GRanges.}
\usage{
easyLoad_narrowPeak(file_paths, file_names = NULL)
}
\arguments{
\item{file_paths}{character vector of paths to narrowPeak files.  If named,
those names will be used in output unless overriden by providing file_names.}

\item{file_names}{character vector of names for output list.  If not NULL will
override any existing names for file_paths.  Default is NULL.}
}
\value{
a named list of GRanges loaded from file_paths
}
\description{
easyLoad_narrowPeak takes a character vector of file paths to narrowPeak
files from MACS2 and returns a named list of GRanges.
}
\examples{
np_f = system.file("extdata/test_loading.narrowPeak",
    package = "seqsetvis", mustWork = TRUE)
easyLoad_narrowPeak(np_f, "my_narrowPeak")
}
