% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_intersect.R
\name{easyLoadGRanges}
\alias{easyLoadGRanges}
\title{Try to load a bed-like file and convert it to a GRanges object}
\usage{
easyLoadGRanges(bed_file, output_GRanges = T)
}
\arguments{
\item{bed_file}{A bed-like file}

\item{output_GRanges}{output data.frame before GRanges conversion}
}
\value{
GRanges with metadata columns describing overlap overlap of input grs
}
\description{
Try to load a bed-like file and convert it to a GRanges object
}
\examples{
TODO easyLoadGRanges()
}
