% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bam_fetch.R
\name{fetchBam}
\alias{fetchBam}
\title{fetch a bam file pileup with the ability to consider cross strand correlation}
\usage{
fetchBam(bam_f, qgr, fragLen = NULL, win_size = 50, target_strand = c("*",
  "+", "-")[1], ...)
}
\arguments{
\item{fragLen}{numeric, NULL, or NA.  if numeric, supplied value is used.}

\item{target_strand}{character. if one of "+" or "-", reads are filtered}

\item{...}{passed to ScanBamParam(), can't be which or what.
if NULL, value is calculated with fragLen_calcStranded
if NA, raw bam pileup with no cross strand shift is returned.
accordingly. ignored if any other value.}
}
\description{
fetch a bam file pileup with the ability to consider cross strand correlation
}
