% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bam_fetch.R
\name{fetchWindowedBam}
\alias{fetchWindowedBam}
\title{fetch a windowed version of a bam file, returns GRanges}
\usage{
fetchWindowedBam(bam_f, qgr, fragLen = NULL, win_size = 50,
  target_strand = c("*", "+", "-")[1])
}
\arguments{
\item{fragLen}{numeric, NULL, or NA.  if numeric, supplied value is used.
if NULL, value is calculated with fragLen_calcStranded
if NA, raw bam pileup with no cross strand shift is returned.
accordingly. ignored if any other value.}

\item{target_strand}{character. if one of "+" or "-", reads are filtered}
}
\description{
fetch a windowed version of a bam file, returns GRanges
}
