% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_process_bw.R
\name{fetchWindowedBigwig}
\alias{fetchWindowedBigwig}
\title{Fetch values from a bigwig appropriate for heatmaps etc.}
\usage{
fetchWindowedBigwig(bw_file, qgr, win_size = 50)
}
\arguments{
\item{bw_file}{The bigwig file to read from.}

\item{qgr}{Set of GRanges to query.  For valid results the width of each
interval should be identical and evenly divisible by \code{win_size}.}

\item{win_size}{The window size that evenly divides widths in \code{qgr}.}
}
\value{
A tidy formatted data.table containing fetched values.
}
\description{
\code{fetchWindowedBigwig} Gets values for each region of the query GRanges (\code{qgr}).
Values correspond to the center of each window of size \code{win_size}.  A tidy formatted data.table
object is returned suitable for plotting using ggplots.
}
\details{
if \code{qgr} contains the range chr1:1-100 and \code{win_size} is
10, values from positions chr1 5,15,25...85, and 95 will be retrieved from \code{bw_file}
}
