% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_signal_operations.R
\name{findMaxPos}
\alias{findMaxPos}
\title{findMaxPos}
\usage{
findMaxPos(prof_dt, qgr, x_ = "x", y_ = "y", by_ = "id", width = 1)
}
\arguments{
\item{prof_dt}{a GRanges or data.table as returned by ssvFetch*.}

\item{qgr}{the GRanges used to query ssvFetch* as the qgr argument.}

\item{x_}{positional variable.  Should almost always be the default, "x".}

\item{y_}{the signal value variable.  Likely the default value of "y" but
could be "y_norm" if append_ynorm was applied to data.}

\item{by_}{region identifier variable. Should almost always be the default,
"id".}

\item{width}{Desired width of final regions.  Default is 1.}
}
\value{
data.table of relative x position from center per id
}
\description{
findMaxPos
}
\examples{
findMaxPos(CTCF_in_10a_profiles_dt, CTCF_in_10a_overlaps_gr)
findMaxPos(CTCF_in_10a_profiles_gr, CTCF_in_10a_overlaps_gr)
}
