% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{fixGRangesWidth}
\alias{fixGRangesWidth}
\title{Derive a new GRanges of consistent width.
Width is selected by rounding up to the lowest multiple of win_size greater than
min_quantile quantile of widths.}
\usage{
fixGRangesWidth(qgr, min_quantile = 0.75, win_size = 1)
}
\arguments{
\item{qgr}{GRanges. To be resized.}

\item{min_quantile}{numeric [0,1]. The quantile level final width must be
greater than. default is 0.75}

\item{win_size}{integer > 0.  final width must be a multiple of win_size.}
}
\value{
a GRanges derived from qgr (length and order match).  All ranges
are of same width and centered on old.  Width is at least minimum quantile
and a multiple of win_size
}
\description{
Derive a new GRanges of consistent width.
Width is selected by rounding up to the lowest multiple of win_size greater than
min_quantile quantile of widths.
}
