% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bam_fetch.R
\name{fragLen_fromMacs2Xls}
\alias{fragLen_fromMacs2Xls}
\title{parse fragLen from MACS2 output}
\usage{
fragLen_fromMacs2Xls(macs2xls_file)
}
\arguments{
\item{macs2xls_file}{character.  an xls file output by MACS2 to parse
frag length from}
}
\value{
numeric fragment length
}
\description{
parse fragLen from MACS2 output
}
\examples{
xls_file = system.file("extdata/test_peaks.xls",
    package = "seqsetvis")
fragLen_fromMacs2Xls(xls_file)
}
