% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{get_mapped_reads}
\alias{get_mapped_reads}
\title{get_mapped_reads}
\usage{
get_mapped_reads(bam_files)
}
\arguments{
\item{bam_files}{Path to 1 or more bam files.  Must be indexed.}
}
\value{
the total mapped reads in each bam file as a named numeric vector.
}
\description{
get_mapped_reads
}
\examples{
bam_file = system.file("extdata/test.bam", package = "seqsetvis", mustWork = TRUE)
get_mapped_reads(bam_file)
}
