% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_plot_functions.R
\name{ggEuler}
\alias{ggEuler}
\title{Try to load a bed-like file and convert it to a GRanges object}
\usage{
ggEuler(memb, line_width = 2)
}
\arguments{
\item{memb}{A membership table}
}
\value{
ggplot of venneuler results
}
\description{
Try to load a bed-like file and convert it to a GRanges object
}
