% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fetch_bam.R
\name{harmonize_seqlengths}
\alias{harmonize_seqlengths}
\title{harmonize_seqlengths}
\usage{
harmonize_seqlengths(gr, bam_file)
}
\arguments{
\item{gr}{GRanges, object to harmonize seqlengths for}

\item{bam_file}{character, a path to a valid bam file}
}
\value{
gr with seqlengths matching bam_file
}
\description{
ensures compatibility between seqlength of gr and bam_file based on header
}
\examples{
library(GenomicRanges)
gr = GRanges("chr1", IRanges(1, 100))
#seqlengths has not been set
seqlengths(gr)
bam = system.file("extdata/test.bam", package = "seqsetvis")
gr2 = harmonize_seqlengths(gr, bam)
#seqlengths now set
seqlengths(gr2)
}
