% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakvisr.R
\docType{package}
\name{peakvisr-package}
\alias{peakvisr}
\alias{peakvisr-package}
\title{Generate R documentation from inline comments.
TESTING
Roxygen2 allows you to write documentation in comment blocks co-located
with code.}
\description{
The only function you're likely to need from \pkg{roxygen2} is
\code{\link{roxygenize}}. Otherwise refer to the vignettes to see
how to format the documentation.
}
