% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions_regionSet.R
\name{regionSetPlotHeatmap}
\alias{regionSetPlotHeatmap}
\title{heatmap}
\usage{
regionSetPlotHeatmap(bw_dt, max_disp = 2000)
}
\arguments{
\item{bw_dt}{data.table of signals}

\item{max_disp}{numeric to sample number of rows down to, default is 2000}
}
\value{
ggplot heatmap of signal profiles, facetted by sample
}
\description{
heatmap
}
