% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqsetvis.R
\docType{package}
\name{seqsetvis-package}
\alias{seqsetvis}
\alias{seqsetvis-package}
\title{easy awesome peak set vis
TESTING
seqsetvis allows you to...}
\description{
2 steps
\code{\link{ssvOverlapIntervalSets}}.
\code{\link{ssvFetchBigwig}}.
Otherwise refer to the vignettes to see
}
\author{
\strong{Maintainer}: Joseph R Boyd \email{jrboyd@uvm.edu} (\href{https://orcid.org/0000-0002-8969-9676}{ORCID})

Authors:
\itemize{
  \item Joseph R Boyd \email{jrboyd@uvm.edu} (\href{https://orcid.org/0000-0002-8969-9676}{ORCID})
}

}
