% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{set_list2memb}
\alias{set_list2memb}
\title{convert a list of sets, each list item should be a character vector
denoting items in sets}
\usage{
set_list2memb(set_list)
}
\arguments{
\item{set_list}{a list of character vectors.  default names will be added if missing}
}
\description{
convert a list of sets, each list item should be a character vector
denoting items in sets
}
