% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fetch_signal.R
\name{setGRangesWidth}
\alias{setGRangesWidth}
\title{Return GRanges with single width}
\usage{
setGRangesWidth(qgr, fwidth, anchor = c("center", "start")[1])
}
\arguments{
\item{qgr}{GRanges}

\item{fwidth}{numeric width to apply}

\item{anchor}{extend from center of from start (strand sensitive)}
}
\value{
GRanges that parallels qgr with fwidth applied according to anchor.
}
\description{
Essentially works like GenomicRanges::resize() but repeated applications
of center do not cause rounding induced drift.
}
