% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions_basicSet.R
\name{setPlotBars2}
\alias{setPlotBars2}
\title{bar plots of set sizes, does fancy facetting}
\usage{
setPlotBars2(object, peak_gr, flip_group_stage = F)
}
\arguments{
\item{object}{passed to setPlotMakeMT for conversion to membership table}

\item{peak_gr}{peak sets before overlapping}

\item{flip_group_stage}{logical, different facetting.  default is FALSE}
}
\value{
ggplot of bar plot of set sizes
}
\description{
bar plots of set sizes, does fancy facetting
}
