% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions_basicSet.R
\name{setPlotEuler}
\alias{setPlotEuler}
\title{Try to load a bed-like file and convert it to a GRanges object}
\usage{
setPlotEuler(object, line_width = 2, shape = c("circle", "ellipse")[1],
  n = 200, fill_circles = T, alpha = 0.15, col_scale = NULL)
}
\arguments{
\item{object}{A membership table}

\item{line_width}{numeric, passed to size aesthetic to control line width}

\item{shape}{shape argument passed to eulerr::euler}

\item{n}{number of points to use for drawing ellipses, passed to  eulerr:::ellipse}

\item{fill_circles}{logical, should circles be filled or empty?}

\item{alpha}{numeric [0,1], alpha value for circle fill}

\item{col_scale}{colors to choose from for circles.  passed to ggplot2 color scales.}
}
\value{
ggplot of venneuler results
}
\description{
Try to load a bed-like file and convert it to a GRanges object
}
\examples{
setPlotEuler(list(1:3, 2:6))
setPlotEuler(CTCF_in_10a_overlaps_gr)
setPlotEuler(S4Vectors::mcols(CTCF_in_10a_overlaps_gr)[,2:3])
}
