% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_feature_plots.R
\name{setPlotHeatmap}
\alias{setPlotHeatmap}
\title{binary heatmap indicating membership.
heatmap is sorted by column left to right.  change column order to reveal patterns}
\usage{
setPlotHeatmap(object, raster_approximation = TRUE, raster_width_min = 1000,
  raster_height_min = 1000)
}
\arguments{
\item{object}{passed to ssvMakeMembTable}

\item{raster_approximation}{instead of standard plot, write temporary raster image and redraw that as plot background.}

\item{raster_width_min}{raster width will be minimun multiple of number of columns over this number}

\item{raster_height_min}{raster height will be minimun multiple of number of rows over this number}
}
\description{
binary heatmap indicating membership.
heatmap is sorted by column left to right.  change column order to reveal patterns
}
