% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPlotMakeMT.R
\docType{methods}
\name{setPlotMakeMT,data.frame-method}
\alias{setPlotMakeMT,data.frame-method}
\title{data.frame input, final output
The final method for most inputs, checks column names}
\usage{
\S4method{setPlotMakeMT}{data.frame}(object)
}
\arguments{
\item{object}{a data.frame}
}
\description{
data.frame input, final output
The final method for most inputs, checks column names
}
\examples{
memb_df = data.frame(a = c(T, T, F, F), b = c(T, F, T, F))
setPlotMakeMT(memb_df)
}
