% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPlotMakeMT.R
\docType{methods}
\name{setPlotMakeMT,DataFrame-method}
\alias{setPlotMakeMT,DataFrame-method}
\title{DataFrame input
setOldClass("DataFrame")}
\usage{
\S4method{setPlotMakeMT}{DataFrame}(object)
}
\arguments{
\item{object}{A single DataFrame (typically from GRanges metadata), will use logical columns}
}
\description{
DataFrame input
setOldClass("DataFrame")
}
