% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPlotMakeMT.R
\docType{methods}
\name{setPlotMakeMT,GRanges-method}
\alias{setPlotMakeMT,GRanges-method}
\title{mcols from GRange input
setOldClass("GRanges")}
\usage{
\S4method{setPlotMakeMT}{GRanges}(object)
}
\arguments{
\item{object}{A single GRanges, will use logical metadata columns}
}
\description{
mcols from GRange input
setOldClass("GRanges")
}
