% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPlotMakeMT.R
\docType{methods}
\name{setPlotMakeMT,GRangesList-method}
\alias{setPlotMakeMT,GRangesList-method}
\title{GRangesList input
setOldClass("GRangesList")}
\usage{
\S4method{setPlotMakeMT}{GRangesList}(object)
}
\arguments{
\item{object}{a list of GRanges, sent to overlapIntervalSets}
}
\description{
GRangesList input
setOldClass("GRangesList")
}
