% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPlotMakeMT.R
\docType{methods}
\name{setPlotMakeMT,list-method}
\alias{setPlotMakeMT,list-method}
\title{list of character vectors input}
\usage{
\S4method{setPlotMakeMT}{list}(object)
}
\arguments{
\item{object}{a list of items that are valid for as.character
A list of GRanges are a special case and will be handled as a GRangesList}
}
\description{
list of character vectors input
}
