% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPlotMakeMT.R
\docType{methods}
\name{setPlotMakeMT,matrix-method}
\alias{setPlotMakeMT,matrix-method}
\title{matrix input}
\usage{
\S4method{setPlotMakeMT}{matrix}(object)
}
\arguments{
\item{object}{A matrix, sent to data.frame to handle column naming}
}
\description{
matrix input
}
\examples{
memb_mat =  matrix(c(T, T, F, F, T, F, T, F), ncol = 2, byrow = F)
setPlotMakeMT(memb_mat)
}
