% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPlotMakeMT.R
\name{setPlotMakeMT}
\alias{setPlotMakeMT}
\title{generic for methods to convert various objects to a logical matrix indicating
membership of items (rows) in sets (columns)}
\usage{
setPlotMakeMT(object)
}
\arguments{
\item{object}{the object to convert}
}
\value{
a logical matrix indicating membership of items (rows) in sets (columns)
}
\description{
generic for methods to convert various objects to a logical matrix indicating
membership of items (rows) in sets (columns)
}
