% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fetch_signal.R
\name{shift_anchor}
\alias{shift_anchor}
\title{orients the relative position of x's zero value and
extends ranges to be contiguous}
\usage{
shift_anchor(score_dt, window_size, anchor)
}
\arguments{
\item{score_dt}{data.table, GRanges() sufficient}

\item{window_size}{numeric, window size used to generate score_dt}

\item{anchor}{character, one of c("center", "center_unstranded",
"left", "left_unstranded")}
}
\value{
score_dt with x values shifted appropriately and start and end
extended to make ranges contiguous
}
\description{
orients the relative position of x's zero value and
extends ranges to be contiguous
}
