% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fetch_signal.R
\name{shift_x0}
\alias{shift_x0}
\title{orients the relative position of x's zero value and
extends ranges to be contiguous}
\usage{
shift_x0(score_dt, window_size, x0)
}
\arguments{
\item{score_dt}{data.table, GRanges() sufficient}

\item{window_size}{numeric, window size used to generate socre_dt}

\item{x0}{character, one of c("center", "center_unstranded",
"left", "left_unstranded")}
}
\value{
score_dt with x values shifted appropriately and start and end
extended to make ranges contiguous
}
\description{
orients the relative position of x's zero value and
extends ranges to be contiguous
}
