% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{ssv_mclapply}
\alias{ssv_mclapply}
\title{ssv_mclapply}
\usage{
ssv_mclapply(X, FUN, mc.cores = getOption("mc.cores", 1), ...)
}
\arguments{
\item{X}{For pbsapply and pblapply, a vector (atomic or list) or an
expressions vector (other objects including classed objects will be coerced
by as.list.) For pbapply an array, including a matrix. For pbtapply an R
object for which a split method exists. Typically vector-like, allowing
subsetting with "[".}

\item{FUN}{The function to be applied to each element of X: see apply,
sapply, and lapply. In the case of functions like +, '%*%', etc., the
function name must be backquoted or quoted. If FUN is NULL, pbtapply
returns a vector which can be used to subscript the multi-way array
pbtapply normally produces.}

\item{mc.cores}{Number of cores to use for pbmclapply. Defaults to option
mc.cores.}

\item{...}{passed to pbapply::pblapply or pbmcapply::pbmclapply}
}
\value{
result of either pblapply or pbmclapply
}
\description{
ssv_mclapply
}
