% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvMakeMembTable.R
\name{ssvFactorizeMembTable}
\alias{ssvFactorizeMembTable}
\title{Convert any object accepted by ssvMakeMembTable to a factor}
\usage{
ssvFactorizeMembTable(object)
}
\arguments{
\item{object}{a valid object for conversion to a membership table by}
}
\value{
a factor representation of object
}
\description{
see \code{\link{ssvMakeMembTable}}
}
