% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_signal_plots.R
\name{ssvSignalBandedQuantiles}
\alias{ssvSignalBandedQuantiles}
\title{plot profiles from bigwigs}
\usage{
ssvSignalBandedQuantiles(bw_dt, y_ = "y", x_ = "x", by_ = "fake",
  hsv_reverse = FALSE, hsv_saturation = 1, hsv_value = 1,
  hsv_grayscale = FALSE, hsv_hue_min = 0, hsv_hue_max = 0.7,
  hsv_symmetric = FALSE, n_quantile = 18, quantile_min = 0.05,
  quantile_max = 0.95)
}
\arguments{
\item{bw_dt}{data.table of bigwig signal}

\item{y_}{the variable name in bw_dt for y axis in plot}

\item{x_}{the variable name in bw_dt for x axis in plot}

\item{by_}{the variable name in bw_dt to facet on}

\item{hsv_reverse}{logical, should color scale be reversed? default FALSE}

\item{hsv_saturation}{numeric [0, 1] saturation for color scale. default 1}

\item{hsv_value}{numeric [0, 1] value for color scale. default 1}

\item{hsv_grayscale}{logical, if TRUE gray() is used instead of rainbow(). default FALSE}

\item{hsv_hue_min}{numeric [0, hsv_hue_max) hue min of color scale}

\item{hsv_hue_max}{numeric (hsv_hue_min, 1] hue max of color scale}

\item{hsv_symmetric}{if TRUE, colorscale is symmetrical, default FALSE.}

\item{n_quantile}{number of evenly size quantile bins}

\item{quantile_min}{the lowest quantile start}

\item{quantile_max}{the highest quantile end}
}
\value{
ggplot object using ribbon plots to show quantile distributions
}
\description{
plot profiles from bigwigs
}
