% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_signal_plots.R
\name{ssvSignalClustering}
\alias{ssvSignalClustering}
\title{clustering as for a heatmap}
\usage{
ssvSignalClustering(bw_dt, nclust = 6, row_ = "id", column_ = "x",
  fill_ = "y", facet_ = "sample", cluster_ = "cluster_id",
  max_rows = 500, max_cols = 100, clustering_col_min = -Inf,
  clustering_col_max = Inf)
}
\arguments{
\item{bw_dt}{data.table of signals}

\item{nclust}{number of clusters}

\item{row_}{variable name mapped to row, likely peak id or gene name for ngs data}

\item{column_}{varaible mapped to column, likely bp position for ngs data}

\item{fill_}{numeric variable to map to fill}

\item{facet_}{variable name to facet horizontally by}

\item{cluster_}{variable name to use for cluster info}

\item{max_rows}{for speed rows are sampled to 500 by default, use Inf to plot full data}

\item{max_cols}{for speed columns are sampled to 100 by default, use Inf to plot full data}

\item{clustering_col_min}{numeric minimum for col range considered when clustering, default in -Inf}

\item{clustering_col_max}{numeric maximum for col range considered when clustering, default in Inf}
}
\value{
data.table of signal profiles, ready for ssvSignalHeatmap
}
\description{
clustering as for a heatmap
}
\examples{
clust_dt = ssvSignalClustering(CTCF_in_10a_profiles_dt)
ssvSignalHeatmap(clust_dt)

clust_dt2 = ssvSignalClustering(CTCF_in_10a_profiles_dt, nclust = 2)
ssvSignalHeatmap(clust_dt2)

#clustering can be targetted to a specific part of the region
clust_dt3 = ssvSignalClustering(CTCF_in_10a_profiles_dt, nclust = 2,
    clustering_col_min = -250, clustering_col_max = -150)
ssvSignalHeatmap(clust_dt3)
clust_dt4 = ssvSignalClustering(CTCF_in_10a_profiles_dt, nclust = 2,
    clustering_col_min = 150, clustering_col_max = 250)
ssvSignalHeatmap(clust_dt4)
}
