% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_CTCF.R
\docType{data}
\name{test_peaks}
\alias{test_peaks}
\title{4 random peaks for single-end data and 4 control regions 30kb downstream from
each peak.}
\format{GRanges length 8}
\description{
matches \code{system.file("extdata/test_peaks.bam", package =
"seqsetvis")}
}
\details{
this is included only for testing ssvFetchBam functions.
}
\keyword{datasets}
