% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fetch_signal.R
\name{viewGRangesWindowed_dt}
\alias{viewGRangesWindowed_dt}
\title{get a windowed sampling of score_gr}
\usage{
viewGRangesWindowed_dt(score_gr, qgr, window_size, x0 = c("center",
  "center_unstranded", "left", "left_unstranded")[1])
}
\arguments{
\item{score_gr}{GRanges with a "score" metadata columns.}

\item{qgr}{regions to view by window.}

\item{window_size}{qgr will be represented by value from score_gr every
window_size bp.}

\item{x0}{character. controls how x value is derived from position for
each region in qgr.  0 may be the left side or center.  If not unstranded,
x coordinates are flipped for (-) strand.}
}
\value{
data.table that is GRanges compatible
}
\description{
Summarizes score_gr by grabbing value of "score" every window_size bp.
Columns in output data.table are:
standard GRanges columns: seqnames, start, end, width, strand
id - matched to names(score_gr). if names(score_gr) is missing,
added as 1:length(score_gr)
y - value of score from score_gr
x - relative bp position
}
\examples{
bam_file = system.file("extdata/test.bam",
    package = "seqsetvis")
qgr = CTCF_in_10a_overlaps_gr[1:5]
qgr = GenomicRanges::resize(qgr, width = 500, fix = "center")
bam_gr = fetchBam(bam_file, qgr)
bam_dt = viewGRangesWindowed_dt(bam_gr, qgr, 50)

if(Sys.info()['sysname'] != "Windows"){
    bw_file = system.file("extdata/MCF10A_CTCF_FE_random100.bw",
        package = "seqsetvis")
    bw_gr = rtracklayer::import.bw(bw_file, which = qgr)
    bw_dt = viewGRangesWindowed_dt(bw_gr, qgr, 50)
}
}
