% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bam_fetch.R
\name{windowViewGRanges_dt}
\alias{windowViewGRanges_dt}
\title{get a windowed sampling of score_gr}
\usage{
windowViewGRanges_dt(score_gr, query_gr, window_size)
}
\description{
get a windowed sampling of score_gr
}
