% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clusteringKmeans.R
\name{within_clust_sort}
\alias{within_clust_sort}
\title{Title
TODO refactor shared code with clusteringKmeansNestedHclust}
\usage{
within_clust_sort(
  clust_dt,
  row_ = "id",
  column_ = "x",
  fill_ = "y",
  facet_ = "sample",
  cluster_ = "cluster_id",
  within_order_strategy = c("hclust", "sort")[2],
  clustering_col_min = -Inf,
  clustering_col_max = Inf,
  dcast_fill = NA
)
}
\arguments{
\item{dcast_fill}{}
}
\value{

}
\description{
Title
TODO refactor shared code with clusteringKmeansNestedHclust
}
