% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_process_bw.R
\name{applySpline}
\alias{applySpline}
\title{applies a spline smoothing to a tidy data.table containing x and y values.}
\usage{
applySpline(dt, x_ = "x", y_ = "y", by_ = "", n = 8, ...)
}
\arguments{
\item{dt}{a tidy data.table containing two-dimensional data}

\item{x_}{the variable name of the x-values}

\item{y_}{the variable name of the y-values}

\item{by_}{optionally, any variables that provide grouping to the data. default is none. see details.}

\item{n}{the number of interpolation points to use per input point, see \code{?spline}}
}
\value{
a newly derived data.table that is \code{n} times longer than original.
}
\description{
\code{applySpline} Is intended for two-dimensional tidy data.tables, as retured by \code{fetchWindowedBigwig}
}
\details{
by_ is quite powerful.  If \code{by_ = c('gene_id', 'sample_id')}, splines
will be calculated individually for each gene in each sample. alternatively if \code{by_ = c('gene_id')}
}
\seealso{
\code{\link{fetchWindowedBigwig}}
}
