% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_process_bw.R
\name{centerAtMax}
\alias{centerAtMax}
\title{centers profile of x and y.  default is to center by region but across all samples.}
\usage{
centerAtMax(dt, x_ = "x", y_ = "y", by_ = NULL, view_size = NULL,
  trim_to_valid = T, check_by_dupes = T)
}
\arguments{
\item{x_}{the variable name of the x-values. default is 'x'}

\item{y_}{the variable name of the y-values default is 'y'}

\item{by_}{optionally, any variables that provide grouping to the data. default is none.  see details.}

\item{view_size}{the size in \code{x_} to consider for finding the max of \code{y_}.
default value of NULL uses complete range of x.
if length(view_size) == 1, range will be c(-view_size, view_size).
if length(view_size) > 1, range will be range(view_size).}
}
\description{
\code{centerAtMax} locates the coordinate x of the maximum in y and shifts x such that it is zero at max y.
}
\details{
by_ is quite powerful.  If \code{by_ = c('gene_id', 'sample_id')}, splines
will be calculated individually for each gene in each sample. alternatively if \code{by_ = c('gene_id')}
}
