% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_signal_plots.R
\name{clusteringKmeans}
\alias{clusteringKmeans}
\title{perform kmeans clustering on matrix rows and return reordered matrix along with order matched cluster assignments
clusters are sorted using hclust on centers}
\usage{
clusteringKmeans(mat, nclust, seed = 0)
}
\arguments{
\item{mat}{numeric matrix to clutser}

\item{nclust}{the number of clusters}

\item{seed}{passed to set.seed() to allow reproducibility}
}
\description{
perform kmeans clustering on matrix rows and return reordered matrix along with order matched cluster assignments
clusters are sorted using hclust on centers
}
