% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{CTCF_in_10a_profiles_dt}
\alias{CTCF_in_10a_profiles_dt}
\title{Profiles for 100 randomly selected regions from overlapping CTCF peaks in 10a cell ChIP-seq
Results from CTCF_in_10a_overlaps_gr}
\format{A tidy data.table of 2100 rows and 9 columns}
\description{
A tidy data.table at window size 100 bp within 300 bp of peak center
 The variables are as follows:
}
\details{
\code{\link{CTCF_in_10a_overlaps_gr CTCF_in_10a_narrowPeak_urls}}

\link{CTCF_in_10a_narrowPeak_urls}

\link{CTCF_in_10a_bigWig_urls}
\enumerate{
  \item seqnames. chromosome for GRanges compatibility
  \item start. start of interval
  \item end. end of interval
  \item width. width of interval
  \item strand. leftover from GRanges.
  \item id. unique identifier
  \item y. fold-enrichment over input.
  \item x. bp relative to center
  \item sample. name of originating sample
}
}
\keyword{datasets}
