% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_CTCF.R
\docType{data}
\name{CTCF_in_10a_profiles_dt}
\alias{CTCF_in_10a_profiles_dt}
\title{Profiles for 100 randomly selected regions from overlapping CTCF peaks in
10a cell ChIP-seq
Results from fetching bigwigs with CTCF_in_10a_overlaps_gr.}
\format{
A tidy data.table of 2100 rows and 9 columns
}
\description{
A tidy data.table at window size 50 bp within 350 bp of peak center
 The variables are as follows:
}
\details{
part of \code{\link{CTCF_in_10a_data}}
\enumerate{
  \item seqnames. chromosome for GRanges compatibility
  \item start. start of interval
  \item end. end of interval
  \item width. width of interval
  \item strand. leftover from GRanges.
  \item id. unique identifier
  \item y. fold-enrichment over input.
  \item x. bp relative to center
  \item sample. name of originating sample
}
}
\keyword{datasets}
