% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fetch_bam.R
\name{fetchWindowedBam}
\alias{fetchWindowedBam}
\title{fetch a windowed version of a bam file, returns GRanges}
\usage{
fetchWindowedBam(bam_f, qgr, fragLen = NULL, win_size = 50,
  target_strand = c("*", "+", "-")[1], return_data.table = FALSE)
}
\arguments{
\item{bam_f}{character or BamFile to load}

\item{qgr}{GRanges regions to fetchs}

\item{fragLen}{numeric, NULL, or NA.  if numeric, supplied value is used.
if NULL, value is calculated with fragLen_calcStranded
if NA, raw bam pileup with no cross strand shift is returned.}

\item{win_size}{numeric >=1.  pileup grabbed every win_size bp}

\item{target_strand}{character. if one of "+" or "-", reads are filtered
accordingly. ignored if any other value.}
}
\value{
tidy GRanges with pileups from bam file.  pileup is
calculated only every win_size bp.
}
\description{
fetch a windowed version of a bam file, returns GRanges
}
\examples{
bam_file = system.file("extdata/test.bam",
    package = "seqsetvis")
qgr = CTCF_in_10a_overlaps_gr[1:5]
bam_gr = fetchWindowedBam(bam_file, qgr)
bam_gr = fetchWindowedBam(bam_file, qgr, fragLen = 180,
    win_size = 10, target_strand = "+")

bam_dt = fetchWindowedBam(bam_file, qgr,
    return_data.table = TRUE)
}
