% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bam_fetch.R
\name{fetchWindowedBam}
\alias{fetchWindowedBam}
\title{fetch a windowed version of a bam file, returns GRanges}
\usage{
fetchWindowedBam(bam_f, qgr, fragLen = NULL, win_size = 50,
  target_strand = c("*", "+", "-")[1])
}
\arguments{
\item{bam_f}{character or BamFile to load}

\item{qgr}{GRanges regions to fetchs}

\item{fragLen}{numeric, NULL, or NA.  if numeric, supplied value is used.
if NULL, value is calculated with fragLen_calcStranded
if NA, raw bam pileup with no cross strand shift is returned.}

\item{win_size}{numeric >=1.  pileup grabbed every win_size bp}

\item{target_strand}{character. if one of "+" or "-", reads are filtered
accordingly. ignored if any other value.}
}
\value{
tidy GRanges with pileups from bam file.  pileup is
calculated only every win_size bp.
}
\description{
fetch a windowed version of a bam file, returns GRanges
}
