% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fetch_bam.R
\name{fetchWindowedBam}
\alias{fetchWindowedBam}
\title{fetch a windowed version of a bam file, returns GRanges}
\usage{
fetchWindowedBam(bam_f, qgr, win_size = 50, win_method = c("sample",
  "summary")[1], summary_FUN = weighted.mean, fragLen = NULL,
  target_strand = c("*", "+", "-")[1], return_data.table = FALSE)
}
\arguments{
\item{bam_f}{character or BamFile to load}

\item{qgr}{GRanges regions to fetchs}

\item{win_size}{numeric >=1.  pileup grabbed every win_size bp}

\item{win_method}{character.  one of c("sample", "summary").  Determines
if \code{\link{viewGRangesWinSample_dt}} or
\code{\link{viewGRangesWinSummary_dt}} is used to represent each region in
qgr.}

\item{summary_FUN}{function.  only relevant if win_method is "summary".
passed to \code{\link{viewGRangesWinSummary_dt}}.}

\item{fragLen}{numeric, NULL, or NA.  if numeric, supplied value is used.
if NULL, value is calculated with fragLen_calcStranded
if NA, raw bam pileup with no cross strand shift is returned.}

\item{target_strand}{character. if one of "+" or "-", reads are filtered
accordingly. ignored if any other value.}

\item{return_data.table}{logical. If TRUE the internal data.table is
returned instead of GRanges.  Default is FALSE.}
}
\value{
tidy GRanges (or data.table if specified) with pileups from bam
file.  pileup is calculated only every win_size bp.
}
\description{
fetch a windowed version of a bam file, returns GRanges
}
\examples{
bam_file = system.file("extdata/test.bam",
    package = "seqsetvis")
qgr = CTCF_in_10a_overlaps_gr[1:5]
bam_gr = fetchWindowedBam(bam_file, qgr)
bam_gr = fetchWindowedBam(bam_file, qgr, fragLen = 180,
    win_size = 10, target_strand = "+")

bam_dt = fetchWindowedBam(bam_file, qgr,
    return_data.table = TRUE)
}
