% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fetch_bigwig.R
\name{fetchWindowedBigwigList}
\alias{fetchWindowedBigwigList}
\title{Iterates a character vector (ideally named) and calls
\code{fetchWindowedBigwig} on each.  Appends grouping variable to each
resulting data.table and uses rbindlist to efficiently combine results.}
\usage{
fetchWindowedBigwigList(file_paths, qgr, unique_names = names(file_paths),
  names_variable = "sample", win_size = 50, x0 = c("left",
  "left_unstranded", "center", "center_unstranded")[3],
  return_data.table = FALSE)
}
\arguments{
\item{file_paths}{The character vector or list of paths to bigwig files to
read from.}

\item{qgr}{Set of GRanges to query.  For valid results the width of each
interval should be identical and evenly divisible by \code{win_size}.}

\item{unique_names}{names to use in final data.table to designate source
bigwig.}

\item{names_variable}{The column name where unique_names are stored.
Default is 'sample'}

\item{win_size}{The window size that evenly divides widths in \code{qgr}.}

\item{x0}{character, one of c("center", "center_unstranded",
"left", "left_unstranded")}

\item{return_data.table}{logical. If TRUE the internal data.table is
returned instead of GRanges.  Default is FALSE.}
}
\value{
A tidy formatted GRanges (or data.table if specified) containing
fetched values.
}
\description{
\code{fetchWindowedBigwigList} iteratively calls \code{fetchWindowedBigwig}.
See \code{\link{fetchWindowedBigwig}} for more info.
}
\details{
if \code{qgr} contains the range chr1:1-100 and \code{win_size} is
10, values from positions chr1 5,15,25...85, and 95 will be
retrieved from \code{bw_file}
}
\examples{
if(Sys.info()['sysname'] != "Windows"){
library(GenomicRanges)
bw_f = system.file("extdata/test_loading.bw",
    package = "seqsetvis", mustWork = TRUE)
bw_files = c("a" = bw_f, "b" = bw_f)
qgr = GRanges("chrTest", IRanges(1, 30))
bw_gr = fetchWindowedBigwigList(bw_files, qgr, win_size = 10)
bw_gr2 = fetchWindowedBigwigList(as.list(bw_files), qgr, win_size = 10)

bw_dt = fetchWindowedBigwigList(bw_files, qgr, win_size = 10,
    return_data.table = TRUE)
}
}
