% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bam_fetch.R
\name{fragLen_calcStranded}
\alias{fragLen_calcStranded}
\title{calculate fragLen from a bam file for specified regions}
\usage{
fragLen_calcStranded(bam_f, qgr, n_regions = 100, force_no_which = FALSE,
  return_fragLen_only = TRUE, max_fragLen = 300)
}
\arguments{
\item{...}{passed to Rsamtools::ScanBamParam, can't be which or what.}
}
\description{
calculate fragLen from a bam file for specified regions
}
