% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions_basicSet.R
\name{setPlotBars}
\alias{setPlotBars}
\title{bar plots of set sizes}
\usage{
setPlotBars(object)
}
\arguments{
\item{object}{passed to setPlotMakeMT for conversion to membership table}
}
\value{
ggplot of bar plot of set sizes
}
\description{
bar plots of set sizes
}
\examples{
setPlotBars(list(1:3, 2:6))
setPlotBars(CTCF_in_10a_overlaps_gr)
setPlotBars(mcols(CTCF_in_10a_overlaps_gr)[,2:3])
}
