% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPlotMakeMT.R
\docType{methods}
\name{setPlotMakeMT}
\alias{setPlotMakeMT}
\alias{setPlotMakeMT,list-method}
\alias{setPlotMakeMT,GRangesList-method}
\alias{setPlotMakeMT,GRanges-method}
\alias{setPlotMakeMT,DataFrame-method}
\alias{setPlotMakeMT,matrix-method}
\alias{setPlotMakeMT,data.frame-method}
\title{generic for methods to convert various objects to a logical matrix indicating
membership of items (rows) in sets (columns)}
\usage{
setPlotMakeMT(object)

\S4method{setPlotMakeMT}{list}(object)

\S4method{setPlotMakeMT}{GRangesList}(object)

\S4method{setPlotMakeMT}{GRanges}(object)

\S4method{setPlotMakeMT}{DataFrame}(object)

\S4method{setPlotMakeMT}{matrix}(object)

\S4method{setPlotMakeMT}{data.frame}(object)
}
\arguments{
\item{object}{the object to convert. Supported types: list (of character or GRanges),
GRanges with membership table metadata, GrangesList,
data.frame/matrix/DataFrame of membership table}
}
\value{
a logical matrix indicating membership of items (rows) in sets (columns)
}
\description{
generic for methods to convert various objects to a logical matrix indicating
membership of items (rows) in sets (columns)

list of character vectors input

GRangesList input
setOldClass("GRangesList")

mcols from GRange input
setOldClass("GRanges")

DataFrame input
setOldClass("DataFrame")

matrix input

data.frame input, final output
The final method for most inputs, checks column names
}
\examples{
char_list = list(letters[1:3], letters[2:4])
setPlotMakeMT(char_list)
library(GenomicRanges)
gr_list = list(GRanges("chr1", IRanges(1:3*2, 1:3*2)), GRanges("chr1", IRanges(2:4*2, 2:4*2)))
setPlotMakeMT(gr_list)
library(GenomicRanges)
gr_list = list(GRanges("chr1", IRanges(1:3*2, 1:3*2)), GRanges("chr1", IRanges(2:4*2, 2:4*2)))
setPlotMakeMT(GRangesList(gr_list))
gr = GRanges("chr1", IRanges(1:3*2, 1:3*2))
gr$set_a = c(TRUE, TRUE, FALSE)
gr$set_b = c(FALSE, TRUE, TRUE)
setPlotMakeMT(gr)
gr = GRanges("chr1", IRanges(1:3*2, 1:3*2))
gr$set_a = c(TRUE, TRUE, FALSE)
gr$set_b = c(FALSE, TRUE, TRUE)
setPlotMakeMT(mcols(gr))
memb_mat =  matrix(c(TRUE, TRUE, FALSE, FALSE, TRUE, FALSE, TRUE, FALSE), ncol = 2, byrow = FALSE)
setPlotMakeMT(memb_mat)
memb_df = data.frame(a = c(TRUE, TRUE, FALSE, FALSE), b = c(TRUE, FALSE, TRUE, FALSE))
setPlotMakeMT(memb_df)
}
