% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions_basicSet.R
\name{setPlotPie}
\alias{setPlotPie}
\title{pie plot of set sizes}
\usage{
setPlotPie(object)
}
\arguments{
\item{object}{object that ssvMakeMembTable can convert to logical matrix membership}
}
\value{
ggplot pie graph of set sizes
}
\description{
pie plot of set sizes
}
\examples{
setPlotPie(list(1:3, 2:6))
setPlotPie(CTCF_in_10a_overlaps_gr)
setPlotPie(S4Vectors::mcols(CTCF_in_10a_overlaps_gr)[,2:3])
}
