% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions_basicSet.R
\name{setPlotPie}
\alias{setPlotPie}
\title{pie plot of set sizes}
\usage{
setPlotPie(object)
}
\arguments{
\item{object}{object that setPlotMakeMT can convert to logical matrix membership}
}
\value{
ggplot pie graph of set sizes
}
\description{
pie plot of set sizes
}
