% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_feature_plots.R
\name{ssvFeatureBars}
\alias{ssvFeatureBars}
\title{bar plots of set sizes}
\usage{
ssvFeatureBars(object, show_counts = TRUE)
}
\arguments{
\item{object}{passed to ssvMakeMembTable for conversion to membership table}

\item{show_counts}{logical.  should counts be displayed at the center of each
bar? default is TRUE}
}
\value{
ggplot of bar plot of set sizes
}
\description{
bar plots of set sizes
}
\examples{
ssvFeatureBars(list(1:3, 2:6))
ssvFeatureBars(CTCF_in_10a_overlaps_gr)
ssvFeatureBars(S4Vectors::mcols(CTCF_in_10a_overlaps_gr)[,2:3])
}
