% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_feature_plots.R
\name{ssvFeatureBinaryHeatmap}
\alias{ssvFeatureBinaryHeatmap}
\title{binary heatmap indicating membership.
heatmap is sorted by column left to right.
change column order to reveal patterns}
\usage{
ssvFeatureBinaryHeatmap(object, raster_approximation = FALSE,
  raster_width_min = 1000, raster_height_min = 1000)
}
\arguments{
\item{object}{passed to ssvMakeMembTable}

\item{raster_approximation}{If TRUE, instead of standard ggplot, write
temporary raster png image and redraw that as plot background. default is
FALSE}

\item{raster_width_min}{raster width will be minimum multiple of number of
columns over this number.  ignored if raster_approximation is FALSE.}

\item{raster_height_min}{raster height will be minimum multiple of number of
rows over this number  ignored if raster_approximation is FALSE}
}
\value{
ggplot using geom_tile of membership table sorted from left to right.
}
\description{
binary heatmap indicating membership.
heatmap is sorted by column left to right.
change column order to reveal patterns
}
\examples{
ssvFeatureBinaryHeatmap(list(1:3, 2:6))
ssvFeatureBinaryHeatmap(CTCF_in_10a_overlaps_gr)
ssvFeatureBinaryHeatmap(S4Vectors::mcols(CTCF_in_10a_overlaps_gr)[,2:3])
ssvFeatureBinaryHeatmap(S4Vectors::mcols(CTCF_in_10a_overlaps_gr)[,3:2])
}
