% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_feature_plots.R
\name{ssvFeatureVenn}
\alias{ssvFeatureVenn}
\title{ggplot implementation of vennDiagram from limma package.  currently limited at 3 sets}
\usage{
ssvFeatureVenn(object, group_names = NULL, counts_txt_size = 5,
  counts_as_labels = FALSE, show_outside_count = FALSE, lwd = 3,
  circle_color = NULL, fill_circles = TRUE,
  fill_alpha = ifelse(fill_circles, 0.5, 0), counts_color = NULL)
}
\arguments{
\item{object}{will be passed to \code{\link{ssvMakeMembTable}} for conversion to membership matrix}

\item{group_names}{useful if names weren't provided or were lost in creating membership matrix}

\item{counts_txt_size}{font size for count numbers}

\item{counts_as_labels}{if TRUE, geom_label is used instead of geom_text.  can be easier to read.}

\item{show_outside_count}{if TRUE, items outside of all sets are counted outside. can be confusing.}

\item{lwd}{uses size aesthetic to control line width of circles.}

\item{circle_color}{colors to use for circle line colors. Uses Dark2 set from RColorBrewer by default.}

\item{fill_circles}{if TRUE, fill circles}

\item{fill_alpha}{alpha value to use for fill, defaults to .5.}

\item{counts_color}{single color to use for displaying counts}
}
\value{
ggplot venn diagram
}
\description{
ggplot implementation of vennDiagram from limma package.  currently limited at 3 sets
}
\examples{
ssvFeatureVenn(list(1:3, 2:6))
ssvFeatureVenn(CTCF_in_10a_overlaps_gr)
ssvFeatureVenn(S4Vectors::mcols(CTCF_in_10a_overlaps_gr)[,2:3])
}
