% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fetch_GRanges.R
\name{ssvFetchGRanges}
\alias{ssvFetchGRanges}
\title{Title}
\usage{
ssvFetchGRanges(grs, qgr, unique_names = names(file_paths),
  win_size = 50, win_method = c("sample", "summary")[1],
  summary_FUN = function(x, w) max(x), target_strand = c("*", "+", "-",
  "both")[1], anchor = c("left", "left_unstranded", "center",
  "center_unstranded")[3], names_variable = "sample",
  return_data.table = FALSE)
}
\arguments{
\item{grs}{a list of GRanges for which to calculate coverage.}

\item{qgr}{Set of GRanges to query.  For valid results the width of each
interval should be identical and evenly divisible by \code{win_size}.}

\item{return_data.table}{}
}
\description{
Title
}
